Edit Roles & Permissions

+++ routes/api.php
@@ -29,22 +29,22 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 //Category routes 
 
-    Route::get('/category' , [CategoryController::class , 'index']);
-    Route::post('/category' , [CategoryController::class , 'store']);
-    Route::get('/category/{category}' , [CategoryController::class , 'show']);
-    Route::patch('/category/{category}' , [CategoryController::class , 'update']);
-    Route::delete('/category/{category}' , [CategoryController::class , 'destroy']);
+    Route::get('/category' , [CategoryController::class , 'index'])->middleware('permission:categories.all');
+    Route::post('/category' , [CategoryController::class , 'store'])->middleware('permission:categories.store');
+    Route::get('/category/{category}' , [CategoryController::class , 'show'])->middleware('permission:categories.show');
+    Route::patch('/category/{category}' , [CategoryController::class , 'update'])->middleware('permission:categories.update');
+    Route::delete('/category/{category}' , [CategoryController::class , 'destroy'])->middleware('permission:categories.delete');
 
 
 /* ---------------------------------------------------------------------------------------------------------------------------------------------*/
 
 //Users routes
 
-    Route::get('/user' , [UserController::class , 'index']);
-    Route::post('/user' , [UserController::class , 'store']);
-    Route::get('/user/{user}' , [UserController::class , 'show']);
-    Route::patch('/user/{user}' , [UserController::class , 'update']);
-    Route::delete('/user/{user}' , [UserController::class , 'destroy']);
+    Route::get('/user' , [UserController::class , 'index'])->middleware('permission:users.all');
+    Route::post('/user' , [UserController::class , 'store'])->middleware('permission:users.store');
+    Route::get('/user/{user}' , [UserController::class , 'show'])->middleware('permission:users.show');
+    Route::patch('/user/{user}' , [UserController::class , 'update'])->middleware('permission:users.update');
+    Route::delete('/user/{user}' , [UserController::class , 'destroy'])->middleware('permission:users.delete');
 
 //Authentication
 
@@ -58,44 +58,44 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 //Factures routes
 
-    Route::get('/facture' , [FactureController::class , 'index']);
-    Route::post('/facture' , [FactureController::class , 'store']);
-    Route::get('/facture/{facture}' , [FactureController::class , 'show']);
-    Route::patch('/facture/{facture}' , [FactureController::class , 'update']);
-    Route::delete('/facture/{facture}' , [FactureController::class , 'destroy']);
+    Route::get('/facture' , [FactureController::class , 'index'])->middleware('permission:factures.all|factures.user');
+    Route::post('/facture' , [FactureController::class , 'store'])->middleware('permission:factures.store');;
+    Route::get('/facture/{facture}' , [FactureController::class , 'show'])->middleware('permission:factures.show');;
+    Route::patch('/facture/{facture}' , [FactureController::class , 'update'])->middleware('permission:factures.update');;
+    Route::delete('/facture/{facture}' , [FactureController::class , 'destroy'])->middleware('permission:factures.delete');;
 
 
 /* ---------------------------------------------------------------------------------------------------------------------------------------------*/
 
 //Opportunities routes
 
-    Route::get('/opportunity' , [OpportunityController::class , 'index']);
-    Route::post('/opportunity' , [OpportunityController::class , 'store']);
-    Route::get('/opportunity/{opportunity}' , [OpportunityController::class , 'show']);
-    Route::patch('/opportunity/{opportunity}' , [OpportunityController::class , 'update']);
-    Route::delete('/opportunity/{opportunity}' , [OpportunityController::class , 'destroy']);
+    Route::get('/opportunity' , [OpportunityController::class , 'index'])->middleware('permission:opportunities.all|opportunities.user');
+    Route::post('/opportunity' , [OpportunityController::class , 'store'])->middleware('permission:opportunities.store');
+    Route::get('/opportunity/{opportunity}' , [OpportunityController::class , 'show'])->middleware('permission:opportunities.show');
+    Route::patch('/opportunity/{opportunity}' , [OpportunityController::class , 'update'])->middleware('permission:opportunities.update');
+    Route::delete('/opportunity/{opportunity}' , [OpportunityController::class , 'destroy'])->middleware('permission:opportunities.delete');
 
 
 /* ---------------------------------------------------------------------------------------------------------------------------------------------*/
 
 //Order routes
 
-    Route::get('/order' , [OrderController::class , 'index']);
-    Route::post('/order' , [OrderController::class , 'store']);
-    Route::get('/order/{order}' , [OrderController::class , 'show']);
-    Route::patch('/order/{order}' , [OrderController::class , 'update']);
-    Route::delete('/order/{order}' , [OrderController::class , 'delete']);
+    Route::get('/order' , [OrderController::class , 'index'])->middleware('permission:orders.all|orders.user');
+    Route::post('/order' , [OrderController::class , 'store'])->middleware('permission:orders.store');
+    Route::get('/order/{order}' , [OrderController::class , 'show'])->middleware('permission:orders.show');
+    Route::patch('/order/{order}' , [OrderController::class , 'update'])->middleware('permission:orders.update');
+    Route::delete('/order/{order}' , [OrderController::class , 'destroy'])->middleware('permission:orders.delete');
 
 
 /* ---------------------------------------------------------------------------------------------------------------------------------------------*/
 
 //Product routes
 
-    Route::get('/product' , [ProductController::class , 'index']);
-    Route::post('/product' , [ProductController::class , 'store']);
-    Route::get('/product/{product}' , [ProductController::class , 'show']);
-    Route::patch('/product/{product}' , [ProductController::class , 'update']);
-    Route::delete('/product/{product}' , [ProductController::class , 'delete']);
+    Route::get('/product' , [ProductController::class , 'index'])->middleware('permission:products.all');
+    Route::post('/product' , [ProductController::class , 'store'])->middleware('permission:products.store');
+    Route::get('/product/{product}' , [ProductController::class , 'show'])->middleware('permission:products.show');
+    Route::patch('/product/{product}' , [ProductController::class , 'update'])->middleware('permission:products.update');
+    Route::delete('/product/{product}' , [ProductController::class , 'destroy'])->middleware('permission:products.delete');
 
 
 /* ---------------------------------------------------------------------------------------------------------------------------------------------*/
@@ -103,11 +103,11 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 //Role routes
 Route::group(['middleware'=>'role:admin'] , function(){
 
-    Route::get('/role' , [RoleController::class , 'index']);
-    Route::post('/role' , [RoleController::class , 'store']);
-    Route::get('/role/{role}' , [RoleController::class , 'show']);
-    Route::patch('/role/{role}' , [RoleController::class , 'update']);
-    Route::delete('/role/{role}' , [RoleController::class , 'delete']);
+    Route::get('/role' , [RoleController::class , 'index'])->middleware('permission:role.all');
+    Route::post('/role' , [RoleController::class , 'store'])->middleware('permission:role.store');
+    Route::get('/role/{role}' , [RoleController::class , 'show'])->middleware('permission:role.show');
+    Route::patch('/role/{role}' , [RoleController::class , 'update'])->middleware('permission:role.update');
+    Route::delete('/role/{role}' , [RoleController::class , 'destroy'])->middleware('permission:role.delete');
     Route::post('/assign', [RoleController::class, 'create']);
 });
